@@ -17,19 +17,20 @@
     "packages/*"
   ],
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "concurrently": "^8.1.0",
-    "esbuild": "0.17.19",
+    "@typescript-eslint/eslint-plugin": "6.13.1",
+    "@typescript-eslint/parser": "6.13.1",
+    "concurrently": "8.2.2",
+    "esbuild": "0.19.8",
     "esbuild-plugin-copy": "2.1.1",
-    "eslint": "^8.42.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
-    "lerna": "6.6.2",
-    "prettier": "^2.8.8",
-    "rimraf": "^5.0.1",
-    "typescript": "^5.0.2"
+    "eslint": "8.55.0",
+    "eslint-config-airbnb-typescript": "17.1.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-import": "2.29.0",
+    "eslint-plugin-prettier": "5.0.1",
+    "lerna": "8.0.0",
+    "prettier": "3.1.0",
+    "rimraf": "5.0.5",
+    "typescript": "5.3.2",
+    "tsx": "3.12.8"
   }
 }
