@@ -2,29 +2,29 @@
   "name": "@mml-playground/mml-doc",
   "private": true,
   "scripts": {
-    "build": "node ./build.js --build",
-    "iterate": "node ./build.js --watch",
+    "build": "tsx ./build.ts --build",
+    "iterate": "tsx ./build.ts --watch",
     "type-check": "tsc --noEmit",
     "lint": "eslint \"./src/**/*.{js,jsx,ts,tsx}\" --max-warnings 0",
     "lint:fix": "eslint \"./src/**/*.{js,jsx,ts,tsx}\" --fix"
   },
   "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
-    "eslint": "^8.43.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-import-resolver-typescript": "3.5.5",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^16.0.1",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "7.32.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "@types/react": "18.2.41",
+    "@types/react-dom": "18.2.17",
+    "eslint": "8.55.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-import-resolver-typescript": "3.6.1",
+    "eslint-plugin-import": "2.29.0",
+    "eslint-plugin-n": "16.3.1",
+    "eslint-plugin-prettier": "5.0.1",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-testing-library": "5.11.0",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1"
+    "eslint-plugin-testing-library": "6.2.0",
+    "@typescript-eslint/eslint-plugin": "6.13.1",
+    "@typescript-eslint/parser": "6.13.1"
   }
 }
