@@ -1,4 +1,4 @@
-import esbuild from "esbuild";
+import * as esbuild from "esbuild";
 import { copy } from "esbuild-plugin-copy";
 
 const buildMode = "--build";
@@ -15,28 +15,32 @@ if (args.length !== 1) {
 
 const mode = args[0];
 
-const buildOptions = {
+const buildOptions: esbuild.BuildOptions = {
   entryPoints: {
     index: "src/index.ts",
   },
   bundle: true,
   write: true,
-  sourcemap: true,
+  sourcemap: "linked",
   outdir: "./build/",
   assetNames: "[dir]/[name]-[hash]",
-  target: "es2020",
+  preserveSymlinks: true,
   loader: {
     ".svg": "file",
     ".png": "file",
     ".jpg": "file",
+    ".glb": "file",
+    ".hdr": "file",
   },
+  outbase: "../",
+  sourceRoot: "./src",
+  publicPath: "/web-client/",
   plugins: [
     copy({
       resolveFrom: "cwd",
       assets: {
         from: ["./public/**/*"],
         to: ["./build/"],
-        keepStructure: true,
       },
     }),
   ],
@@ -48,9 +52,12 @@ switch (mode) {
     break;
   case watchMode:
     esbuild
-      .context({ ...buildOptions, banner: {
+      .context({
+        ...buildOptions,
+        banner: {
           js: ` (() => new WebSocket((window.location.protocol === "https:" ? "wss://" : "ws://")+window.location.host+'/web-client-build').addEventListener('message', () => location.reload()))();`,
-        }, })
+        },
+      })
       .then((context) => context.watch())
       .catch(() => process.exit(1));
     break;
